refactor(mercadopago): Export cloud function in v1 and script to build
