+++ packages/apps/mercadopago/firebase.ts
@@ -1,0 +1,2 @@
+/* eslint-disable */ 
+export * from './lib/firebase'
