+++ packages/events/src/firebase.ts
@@ -3,4 +3,4 @@ import '@cloudcommerce/firebase/lib/init';
 
 export * from '@cloudcommerce/app-tiny-erp';
 
-export { mercadopago } from '@cloudcommerce/app-mercadopago';
+export * from '@cloudcommerce/app-mercadopago/firebase';
